Move default text styles to folder
